+++ testfiles/ids.py
@@ -1,0 +1,8 @@
+class foo:
+    pass
+
+foo = 1
+bar = foo
+
+def bar(baz, bang=1, foo=2):
+    pass